Use dedicated counter project
If we use counter directly from Jepsen test, it's create on the master note,
not on the workers. We need to use dedicated counter project which we build
and push to the workers and then start it there. This is prepare part of this
process when we build and push the counter project.